+++ sain/macros.py
@@ -395,9 +395,9 @@ def assert_eq(left: T, right: T) -> None:
     assert_eq(a, b)
     ```
     """
-    assert (
-        left == right
-    ), f'assertion `left == right` failed\nleft: "{left!r}"\nright: "{right!r}"'
+    assert left == right, (
+        f'assertion `left == right` failed\nleft: "{left!r}"\nright: "{right!r}"'
+    )
 
 
 @rustc_diagnostic_item("assert_ne")
@@ -415,9 +415,9 @@ def assert_ne(left: T, right: T) -> None:
     assert_ne(a, b)
     ```
     """
-    assert (
-        left != right
-    ), f'assertion `left != right` failed\nleft: "{left!r}"\nright: "{right!r}"'
+    assert left != right, (
+        f'assertion `left != right` failed\nleft: "{left!r}"\nright: "{right!r}"'
+    )
 
 
 @rustc_diagnostic_item("include_bytes")